Added the concept of a measurement filter
